添加数值计算模块有关n次多项式、函数插值、函数拟合的单元测试

+++ modules/camera/src/opt_camera/opt_camera.cpp
@@ -241,7 +241,7 @@ bool OptCamera::Impl::retrieve(cv::OutputArray image) noexcept
         return true;
     };
     // retrieve failed
-    ERROR_("opt - Failed to retrieve, unsupported mode: %d.", flag);
+    ERROR_("opt - Failed to retrieve, unsupported mode: %d.", static_cast<int>(flag));
     image.assign(cv::Mat());
     return false;
 }
